pyproject.toml
[project]
name = "kaggle-template"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # データサイエンス基本ライブラリ
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # 機械学習
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    # Jupyter
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    # ユーティリティ
    "kaggle>=1.5.16",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "mypy>=1.5.0",
    "ipython",
]
